[refactor] package 이름 수정 spotUpdater -> spot
