+++ src/main/java/com/example/spring/domain/spot/SpotRepository.java
@@ -1,6 +1,6 @@
-package com.example.spring.domain.spotUpdater;
+package com.example.spring.domain.spot;
 
-import com.example.spring.domain.spotUpdater.domain.Spot;
+import com.example.spring.domain.spot.domain.Spot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SpotRepository extends JpaRepository<Spot,String> {
